% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube1.R, R/square.R
\name{cube1}
\alias{cube1}
\alias{square1}
\title{Raise a vector to a power a specific vector}
\usage{
cube1(x)

square1(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Raise a vector to a power a specific vector
}
\examples{
square1(3)
square1(1:5)
cube1(3)
cube1(1:4)
}
