% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplot.R
\name{quickplot1}
\alias{quickplot1}
\title{Super quick function that creates plots}
\usage{
quickplot1(
  x = gapminder::gapminder$gdpPercap,
  y = gapminder::gapminder$lifeExp,
  data = gapminder::gapminder
)
}
\arguments{
\item{x}{the x axis (default gapminder::gapminder$gdpPercap)}

\item{y}{the y axis (default gapminder::gapminder$lifeExp)}

\item{data}{by default gapminder::gapminder}
}
\value{

}
\description{
Super quick function that creates plots
}
